requirements.txt
audioplayer~=0.6
graphics.py~=5.0.1.post1